nlp_service/requirements.txt
fastapi
uvicorn
pydantic
transformers
torch
python-dotenv
